move loci parsing logic from Common into loci pkg
