+++ src/main/scala/org/hammerlab/guacamole/commands/GermlineAssemblyCaller.scala
@@ -194,7 +194,7 @@ object GermlineAssemblyCaller {
 
     override def run(args: Arguments, sc: SparkContext): Unit = {
       val reference = ReferenceBroadcast(args.referenceFastaPath, sc)
-      val loci = Common.lociFromArguments(args)
+      val loci = args.parseLoci()
       val readSet = Common.loadReadsFromArguments(
         args,
         sc,
