+++ src/main/scala/org/hammerlab/guacamole/commands/SomaticStandardCaller.scala
@@ -69,7 +69,7 @@ object SomaticStandard {
 
     override def run(args: Arguments, sc: SparkContext): Unit = {
       VariantUtils.validateArguments(args)
-      val loci = Common.lociFromArguments(args)
+      val loci = args.parseLoci()
       val filters =
         InputFilters(
           overlapsLoci = loci,
